Overhaul everything to use stuff generated by pip2nix
This was needed because in the previous approach we used some python packages
from nixpkgs and some were manually overrided to have a specific version.
Although this approach had the benefit of following nixpkgs closely, updating
some arbitrary parts of the package hierarchy by hand upon updating
mautrix-hangouts to a newer version got tiresome really quickly. Hence the repo
was entirelly changed to use python package definitions generated by pip2nix.
It's nice because it automatically generates expressions for all of the
transitive dependencies of mautrix-hangouts.